Use 60min window for ping loss
a) If the package loss is more than 90% in that window, the router
   switches in 'problem' state. This corresponds to 54min.

b) When the router is in problem state and comes online again,
   it is recovering. When the calculated package loss in the
   window drops below 30%, the node switches back to 'ok' state.
   This correspondents to the node being online again for 18min.

This is kind of a hotfix, as a lot of maintainance is done recently.
Routers are considered offline, because their ipv6 routing is broken.
But most of the time this maintainance is done by night or only takes
a few minutes.