Subscribe with JS & fetch()
Before this commit, when a user wanted to subscribe to a lot of nodes,
it took a lot of time. The reason for that was, that the /subscribe page
was revisited after each subscription. Unfortunately the /subscribe page
is served only after a couple of seconds, because internally a recent
nodes.json is downloaded and processed. Especially for users who want to
subscribe to a bunch of nodes, this became tedious.

With this commit, javascripts asynchronous fetch() api is used to
subscribe to new nodes. This way the page does not have to be reloaded
for each subscription. If the user has disabled javascript, the page
should work as before.